# user-module/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User module API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/departments": {
            "get": {
                "tags": [
                    "Department"
                ],
                "summary": "Get list departments",
                "operationId": "da8820891dd6bf08713e9d77fbe6389f",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of department",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort by column",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List departments"
                    }
                }
            },
            "post": {
                "tags": [
                    "Department"
                ],
                "summary": "Create new department",
                "operationId": "c0f48115550ccf6a76b43ff263bfc05c",
                "requestBody": {
                    "description": "Create new department",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Department 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Department successfully created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/departments/{id}": {
            "get": {
                "tags": [
                    "Department"
                ],
                "summary": "Get department by id",
                "operationId": "ecdae8bd198987eca46831ced8777bba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of department",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Department"
                    }
                }
            },
            "put": {
                "tags": [
                    "Department"
                ],
                "summary": "Update department",
                "operationId": "1d6fe343b0c809ce2a0722693aed793d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of department",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update department",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Department 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Department successfully updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Department"
                ],
                "summary": "Delete department",
                "operationId": "5cc7aebc708fbbbeb04b81504f62b499",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of department",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Department successfully deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/template/department": {
            "get": {
                "tags": [
                    "Department"
                ],
                "summary": "Get template",
                "operationId": "ddda9c5f7b76b7faf6faff4309dfc657",
                "responses": {
                    "200": {
                        "description": "Template file"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Department"
                ],
                "summary": "Import departments",
                "operationId": "df3af3e91a6d4a4a3c4ce05436e37d40",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "excel file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}